separate calendar api handling from app.js into services/calendar.js
