+++ src/app.js
@@ -21,6 +21,7 @@ const assignmentDB = require('./route-handlers/db-assignments.js');
 const sessionDB = require('./route-handlers/db-sessions.js');
 const participantDB = require('./route-handlers/db-participants.js');
 const cronofy = require('cronofy');
+const calApi = require('./services/calendar.js'); 
  
 const OAuth2 = google.auth.OAuth2;
 const app = express(feathers());
@@ -53,8 +54,6 @@ app.hooks(appHooks);
 // ===============================
 app.post('/login', (req,res) => {
   const user = new OAuth2(process.env.GOOGLE_CALENDAR_CLIENT_ID, process.env.GOOGLE_CALENDAR_SECRET, '');
-  let teacher;
-  let requestedCalendar = [];
   user.verifyIdToken(
     req.body.idtoken,
     process.env.GOOGLE_CALENDAR_CLIENT_ID,
@@ -71,36 +70,12 @@ app.post('/login', (req,res) => {
         });
       
         let client = new cronofy({
-          access_token: 't5cezkIyHkr2pU-w1c9Q6fneYdhyqu_e',
+          access_token: process.env.CRONOFY_ACCESS_TOKEN,
         });
-        var options = {
-          tzid: 'Etc/UTC'
-        };
-
-        client.listCalendars(options)
-          .then(function (response) {
-            console.log('calendars available for list');
-            // var calendars = response.calendars;
-          })
-          .catch(err => {
-            console.log(err);
-          });
-    
-        client.readEvents(options)
-          .then(function (response) {
-            console.log('calendar events available');
-            let teacherCalendar = response.events;
-            for(let i=0; i<teacherCalendar.length; i++){
-              if(teacherCalendar[i].organizer.display_name === '5th Grade'){
-                requestedCalendar.push(teacherCalendar[i]);
-              }
-            }
-            sendTeacher(teacher, requestedCalendar);
-            // console.log('teacher calendar is ', teacherCalendar);
-          })
-          .catch(err => {
-            console.log(err);
-          });
+        
+        //call calendar API for calendar events
+        const calendarName = '5th Grade';
+        calApi.getCalendar(client, calendarName);
 
         userDB.findOrCreateTeacher(userPayload)
           .then((response) => {
@@ -112,11 +87,6 @@ app.post('/login', (req,res) => {
       }
     }
   );
-  const sendTeacher = (teach, calendar) => {
-    console.log('teacher is ', teach);
-    console.log('teacher calendar is ', calendar);
-    res.status(201).send({teacherData: teach, calendarData: calendar});
-  }
 });
 
 app.post('/studentLogin', (req,res) => {
